mdh
Today I have created a new android application. Whose name is Dewaara. This app contains new android feature floating point button, hide button, current year calendar, menu bar, tool bar and search bar etc. Have been brought to the home activity
Working Process
When you open the app, you will see a welcome screen in which you will see Dewaara company logo rounding. Then, after a few seconds, the home activity will open, which will display all the features shown above, one of which will be the switch button, on which the floating point button and the calendar will be hidden and on the off will be a different display and on scroll, the floating bar up- will be down. On the floating point click it will move the right side and after double click again it will come back to the center and in this app the tool bar, menu bar, floating point button etc. Located in the bottom. Which is the new feature of android. Upcoming features will come to this app soon…  Thank you!!!